+++ lib/brando/images/operations/sizing.ex
@@ -54,41 +54,197 @@ defmodule Brando.Images.Operations.Sizing do
     Progress.update_progress(user, "#{filename} — Oppretter bildestørrelse: <strong>#{size_key}</strong>", key: to_string(id) <> size_key)
     File.mkdir_p(image_dest_dir)
 
-    image_src_path = Images.Utils.media_path(image_src)
-    image_dest_path = Images.Utils.media_path(image_dest)
-
-    with true <- File.exists?(image_src_path),
-         image <- Mogrify.open(image_src_path) do
-      quality = Map.get(size_cfg, "quality", "100")
-
-      size_cfg = get_size_cfg_orientation(size_cfg, height, width)
-
-      if size_cfg["crop"] do
-        geometry_with_focal = calculate_geometry_with_focal(focal, width, height, size_cfg)
-
-        image
-        |> Mogrify.resize_to_fill(geometry_with_focal)
-        |> Mogrify.resize_to_fill(size_cfg["size"])
-        |> Mogrify.custom("quality", quality)
-        |> Mogrify.save(path: image_dest_path)
-      else
-        image
-        |> Mogrify.resize_to_limit(size_cfg["size"])
-        |> Mogrify.custom("quality", quality)
-        |> Mogrify.save(path: image_dest_path)
-      end
-
-      Progress.update_progress(user, "#{filename} — Oppretter bildestørrelse: <strong>#{size_key}</strong>", key: to_string(id) <> size_key, percent: 100)
-
-      {:ok, %Images.TransformResult{
+    conversion_parameters =
+      %Images.ConversionParameters{
         id: id,
         size_key: size_key,
-        image_path: image_dest
-      }}
+        image_src_path: Images.Utils.media_path(image_src),
+        image_dest_path: Images.Utils.media_path(image_dest),
+        image_dest_rel_path: image_dest,
+        original_width: width / 1,
+        original_height: height / 1
+      }
+
+    with {:ok, {:image, :exists}} <- image_src_exists(conversion_parameters) do
+      conversion_parameters =
+        conversion_parameters
+        |> add_size_cfg(size_cfg)
+        |> add_quality()
+        |> add_focal_point(focal)
+        |> add_open_image()
+        |> add_crop_flag()
+        |> add_crop_dimensions()
+        |> add_resize_dimensions()
+        |> add_anchor()
+        |> add_geographies()
+      Progress.update_progress(user, "#{filename} — Oppretter bildestørrelse: <strong>#{size_key}</strong>", key: to_string(id) <> size_key, percent: 100)
+      process_image(conversion_parameters)
     else
-      false ->
-        {:error, {:create_image_size, {:file_not_found, image_src_path}}}
+      {:error, {:image, :not_found}} ->
+        {:error, {:create_image_size, {:file_not_found, conversion_parameters.image_src_path}}}
     end
+  end
+
+  def process_image(%Images.ConversionParameters{
+    id: id,
+    size_key: size_key,
+    crop: false,
+    image: image,
+    quality: quality,
+    image_dest_path: image_dest_path,
+    image_dest_rel_path: image_dest_rel_path,
+    resize_geography: resize_geography
+  }) do
+    image
+    |> Mogrify.custom("quality", quality)
+    |> Mogrify.resize_to_limit(resize_geography)
+    |> Mogrify.save(path: image_dest_path)
+
+    {:ok, %Images.TransformResult{
+      id: id,
+      size_key: size_key,
+      image_path: image_dest_rel_path
+    }}
+  end
+
+  def process_image(%Images.ConversionParameters{
+    id: id,
+    size_key: size_key,
+    crop: true,
+    image: image,
+    quality: quality,
+    image_dest_path: image_dest_path,
+    image_dest_rel_path: image_dest_rel_path,
+    resize_geography: resize_geography,
+    crop_geography: crop_geography,
+  }) do
+    image
+    |> Mogrify.custom("quality", quality)
+    |> Mogrify.custom("resize", resize_geography)
+    |> Mogrify.custom("crop", crop_geography)
+    |> Mogrify.save(path: image_dest_path)
+
+    {:ok, %Images.TransformResult{
+      id: id,
+      size_key: size_key,
+      image_path: image_dest_rel_path
+    }}
+  end
+
+  def image_src_exists(%Images.ConversionParameters{image_src_path: src}) do
+    case File.exists?(src) do
+      true -> {:ok, {:image, :exists}}
+      false -> {:error, {:image, :not_found}}
+    end
+  end
+
+  def add_focal_point(conversion_parameters, focal) do
+    focal =
+      focal
+      |> Map.put("x", is_binary(focal["x"]) && String.to_integer(focal["x"]) || focal["x"])
+      |> Map.put("y", is_binary(focal["y"]) && String.to_integer(focal["y"]) || focal["y"])
+
+    Map.put(conversion_parameters, :focal_point, focal)
+  end
+
+  def add_size_cfg(%{original_width: width, original_height: height} = conversion_parameters, size_cfg) do
+    Map.put(conversion_parameters, :size_cfg, get_size_cfg_orientation(size_cfg, height, width))
+  end
+
+  def add_open_image(conversion_parameters) do
+    image = Mogrify.open(conversion_parameters.image_src_path)
+    Map.put(conversion_parameters, :image, image)
+  end
+
+  def add_crop_flag(%{size_cfg: %{"crop" => crop}} = conversion_parameters) do
+    Map.put(conversion_parameters, :crop, crop)
+  end
+
+  def add_crop_flag(%{size_cfg: _} = conversion_parameters) do
+    Map.put(conversion_parameters, :crop, false)
+  end
+
+  def add_quality(%{size_cfg: size_cfg} = conversion_parameters) do
+    quality = Map.get(size_cfg, "quality", "100")
+    Map.put(conversion_parameters, :quality, quality)
+  end
+
+  def add_crop_dimensions(%{crop: true, size_cfg: size_cfg} = conversion_parameters) do
+    {crop_width, crop_height} = get_crop_dimensions_from_cfg(size_cfg)
+
+    conversion_parameters
+    |> Map.put(:crop_width, crop_width / 1)
+    |> Map.put(:crop_height, crop_height / 1)
+  end
+
+  def add_crop_dimensions(%{crop: false} = conversion_parameters) do
+    conversion_parameters
+  end
+
+  @doc """
+  Calculate resize dimension while keeping aspect ratio
+  """
+  def add_resize_dimensions(%{
+    crop: true,
+    crop_width: crop_width,
+    crop_height: crop_height,
+    original_width: original_width,
+    original_height: original_height
+  } = conversion_parameters) do
+    {resize_width, resize_height} =
+      if crop_width > crop_height do
+        resize_width = crop_width
+        resize_height = (crop_width * original_height) / original_width
+
+        if resize_height < crop_height do
+          resize_width = (crop_height * resize_width) / resize_height
+          resize_height = crop_height
+
+          {resize_width, resize_height}
+        else
+          {resize_width, resize_height}
+        end
+      else
+        resize_width = (crop_height * original_width) / original_height
+        resize_height = crop_height
+
+        if resize_width < crop_width do
+          resize_height = (crop_width * resize_height) / resize_width
+          resize_width = crop_width
+
+          {resize_width, resize_height}
+        else
+          {resize_width, resize_height}
+        end
+      end
+
+    conversion_parameters
+    |> Map.put(:resize_width, resize_width / 1)
+    |> Map.put(:resize_height, resize_height / 1)
+  end
+
+  def add_resize_dimensions(%{crop: false} = conversion_parameters) do
+    conversion_parameters
+  end
+
+  def add_geographies(%{
+    crop: true,
+    anchor: anchor,
+    resize_width: resize_width,
+    resize_height: resize_height,
+    crop_width: crop_width,
+    crop_height: crop_height
+  } = conversion_parameters) do
+    resize_geography = "#{resize_width}x#{resize_height}"
+    crop_geography = "#{crop_width}x#{crop_height}+#{anchor.x}+#{anchor.y}"
+
+    conversion_parameters
+    |> Map.put(:resize_geography, resize_geography)
+    |> Map.put(:crop_geography, crop_geography)
+  end
+
+  def add_geographies(%{crop: false, size_cfg: %{"size" => resize_geography}} = conversion_parameters) do
+    Map.put(conversion_parameters, :resize_geography, resize_geography)
   end
 
   defp get_size_cfg_orientation(size_cfg, width, height) do
@@ -103,32 +259,88 @@ defmodule Brando.Images.Operations.Sizing do
     end
   end
 
-  defp calculate_geometry_with_focal(focal, width, height, size_cfg) do
-    focal = %{
-      x: focal["x"] / 100 * width,
-      y: focal["y"] / 100 * height
+  def add_anchor(%{
+    crop: true
+  } = conversion_parameters) do
+    conversion_parameters
+    |> get_original_focal_point()
+    |> transform_focal_point()
+    |> calculate_anchor()
+  end
+
+  def add_anchor(%{crop: false} = conversion_parameters) do
+    conversion_parameters
+  end
+
+  def calculate_anchor(%{
+    transformed_focal_point: transformed_focal_point,
+    resize_width: resize_width,
+    resize_height: resize_height,
+    crop_width: crop_width,
+    crop_height: crop_height
+  } = conversion_parameters) do
+    anchor = %{
+      x: transformed_focal_point.x - (crop_width / 2),
+      y: transformed_focal_point.y - (crop_height / 2)
     }
 
-    [resize_width, resize_height] =
-      size_cfg["size"]
+    anchor = %{
+      x: (anchor.x + crop_width <= resize_width) && anchor.x || resize_width - crop_width,
+      y: (anchor.y + crop_height <= resize_height) && anchor.y || resize_height - crop_height
+    }
+
+    # Ensure that the crop area doesn't fall off the top left of the image.
+    anchor = %{
+      x: max(0, anchor.x),
+      y: max(0, anchor.y),
+    }
+
+    Map.put(conversion_parameters, :anchor, anchor)
+  end
+
+  @doc """
+  Get pixel X and Y of focal point
+  """
+  def get_original_focal_point(%{
+    focal_point: focal,
+    original_width: original_width,
+    original_height: original_height
+  } = conversion_parameters) do
+    original_focal_point = %{
+      x: ((focal["x"] / 1) / 100 * original_width) / 1,
+      y: ((focal["y"] / 1) / 100 * original_height) / 1
+    }
+
+    Map.put(conversion_parameters, :original_focal_point, original_focal_point)
+  end
+
+  @doc """
+  Transform original focal point to a scaled version fitting the resized image dimensions
+  """
+  def transform_focal_point(%{
+    original_focal_point: original_focal_point,
+    original_width: original_width,
+    original_height: original_height,
+    resize_width: resize_width,
+    resize_height: resize_height
+  } = conversion_parameters) do
+    transformed_focal_point = %{
+      x: (((original_focal_point.x / 1) / (original_width / 1)) * resize_width) / 1,
+      y: (((original_focal_point.y / 1) / (original_height / 1)) * resize_height) / 1
+    }
+
+    Map.put(conversion_parameters, :transformed_focal_point, transformed_focal_point)
+  end
+
+  @doc """
+  Extract crop dimensions from image config as a tuple of integers
+  """
+  def get_crop_dimensions_from_cfg(cfg) do
+    [target_width, target_height] =
+      cfg["size"]
       |> String.replace(~r/\^|\!|\>|\<|\%/, "")
       |> String.split("x")
 
-    k = String.to_integer(resize_width) / String.to_integer(resize_height)
-
-    wm = width
-    hm = wm / k
-
-    [hm, wm] =
-      if hm > height do
-        [height, height * k]
-      else
-        [hm, wm]
-      end
-
-    fx2 = focal.x * wm / width
-    fy2 = focal.y * hm / height
-
-    "#{wm}x#{hm}+#{focal.x - fx2}+#{focal.y - fy2}"
+    {String.to_integer(target_width), String.to_integer(target_height)}
   end
 end